+++ es/packages/protoc-gen-es-cybozu-validate/tsconfig.json
@@ -4,5 +4,5 @@
     "outDir": "lib",
     "resolveJsonModule": true
   },
-  "include": ["src/*.ts", "./package.json"]
+  "include": ["src/**/*.ts", "./package.json"]
 }
